chore(nvim)!: change `J` and `K` mappings
I wanted to keep these mappings 'cause it's easier to move around with
them unlike `C-d` and `C-u` which I find to be slightly aggressive. But my
long term goal is to be as close to default neovim as possible to reduce
maintenance burden and easier get-going in server environments.

`K` is mapped to `keywordprg` in default and I wanted to keep that and since
keeping only the `J` mapping doesn't make any sense so I removed them both.

I know it's a pain but will get used to it as the time goes.